Clean up in acceptance module
- Convert some tests to TCK scenarios
- Move left-over tests in small classes to larger ones
- Rename test classes to more accurately reflect contents
- Improve test structure
