--- community/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/ReturnAcceptanceTest.scala
@@ -1,53 +1,0 @@
-/*
- * Copyright (c) 2002-2016 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * This file is part of Neo4j.
- *
- * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.neo4j.internal.cypher.acceptance
-
-import org.neo4j.cypher._
-import org.neo4j.cypher.internal.compiler.v3_1.test_helpers.CustomMatchers
-import org.neo4j.graphdb._
-
-class ReturnAcceptanceTest extends ExecutionEngineFunSuite with CustomMatchers with NewPlannerTestSupport {
-
-  test("should return shortest paths if using a ridiculously unhip cypher") {
-    val a = createNode()
-    val b = createNode()
-    val c = createNode()
-    relate(a, b)
-    relate(b, c)
-
-    val result = eengine.execute("match (a), (c) where id(a) = 0 and id(c) = 2 return shortestPath((a)-[*]->(c))", Map.empty[String, Any], graph.session()).columnAs[Path]("shortestPath((a)-[*]->(c))").toList.head
-    result.endNode() should equal(c)
-    result.startNode() should equal(a)
-    result.length() should equal(2)
-  }
-
-  test("should return shortest path") {
-    val a = createLabeledNode("Start")
-    val b = createNode()
-    val c = createLabeledNode("End")
-    relate(a, b)
-    relate(b, c)
-
-    val result = executeWithAllPlannersAndCompatibilityMode("match (a:Start), (c:End) return shortestPath((a)-[*]->(c))").columnAs[Path]("shortestPath((a)-[*]->(c))").toList.head
-    result.endNode() should equal(c)
-    result.startNode() should equal(a)
-    result.length() should equal(2)
-  }
-}
